feat: Non interactive command with reusable input definitions
