+++ packages/cli/src/plugins/constants.ts
@@ -1,0 +1,2 @@
+export const PLUGIN_OPTIONS = Symbol('PLUGIN_OPTIONS');
+export const loggerCtx = 'Plugin';
