+++ packages/cli/src/plugins/types.ts
@@ -1,0 +1,7 @@
+/**
+ * @description
+ * The plugin can be configured using the following options:
+ */
+export interface PluginInitOptions {
+    exampleOption?: string;
+}
